+++ backend/stout/src/main/java/edu/mtech/stout/resources/ScaleResource.java
@@ -1,8 +1,12 @@
 package edu.mtech.stout.resources;
 
+import edu.mtech.stout.api.QueryBySelector;
 import edu.mtech.stout.api.Status;
 import edu.mtech.stout.core.Scale;
+import edu.mtech.stout.core.User;
+import edu.mtech.stout.db.ProgramDAO;
 import edu.mtech.stout.db.ScaleDAO;
+import io.dropwizard.auth.Auth;
 import io.dropwizard.hibernate.UnitOfWork;
 import io.dropwizard.jersey.PATCH;
 import io.dropwizard.jersey.params.LongParam;
@@ -15,7 +19,8 @@ import javax.ws.rs.core.MediaType;
 @Produces(MediaType.APPLICATION_JSON)
 public class ScaleResource {
 
-  ScaleDAO dao;
+  private ScaleDAO dao;
+  private QueryBySelector queryBySelector = new QueryBySelector();
 
   public ScaleResource(ScaleDAO dao) {
     this.dao = dao;
@@ -23,8 +28,10 @@ public class ScaleResource {
 
   @GET
   @UnitOfWork
-  public Scale getScale(@PathParam("scaleId") LongParam scaleId) {
-    return findSafely(scaleId.get());
+  public Scale getScale(@Auth User user, @PathParam("scaleId") LongParam scaleId) {
+    if(queryBySelector.queryByProgramId(user,findSafely(scaleId.get()).getProgId()))
+      return findSafely(scaleId.get());
+    throw new NotAuthorizedException("Cannot get a scale not in a program you do not have access to");
   }
 
   private Scale findSafely(long scaleId) {
@@ -33,29 +40,36 @@ public class ScaleResource {
 
   @PATCH
   @UnitOfWork
-  public Scale updateScale(@PathParam("scaleId") LongParam scaleId, Scale scale) {
-    return dao.update(scale);
+  public Scale updateScale(@Auth User user, @PathParam("scaleId") LongParam scaleId, Scale scale)
+  {
+    if(queryBySelector.queryByProgramId(user, scale.getProgId()))
+      return dao.update(scale);
+    throw new NotAuthorizedException("Cannot update a scale not in a program you do not have access to");
   }
 
   @DELETE
   @RolesAllowed({"Admin", "Program Coordinator"})
   @UnitOfWork
-  public Status deleteScale(@PathParam("scaleId") LongParam scaleId) {
-    Status status = new Status();
-    status.setId(scaleId.get().intValue());
-    status.setAction("DELETE");
-    status.setResource("Scale");
+  public Status deleteScale(@Auth User user, @PathParam("scaleId") LongParam scaleId) {
+    if(queryBySelector.queryByProgramId(user,findSafely(scaleId.get()).getProgId())) {
+      Status status = new Status();
+      status.setId(scaleId.get().intValue());
+      status.setAction("DELETE");
+      status.setResource("Scale");
 
-    boolean success = dao.delete(findSafely(scaleId.get().intValue()));
+      boolean success = dao.delete(findSafely(scaleId.get().intValue()));
 
-    if (success) {
-      status.setMessage("Successfully deleted scale");
-      status.setStatus(200);
-    } else {
-      status.setMessage("Error deleting scale");
-      status.setStatus(500);
+      if (success) {
+        status.setMessage("Successfully deleted scale");
+        status.setStatus(200);
+      } else {
+        status.setMessage("Error deleting scale");
+        status.setStatus(500);
+      }
+
+      return status;
+    }else{
+      throw new NotAuthorizedException("Cannot delete a scale not in a program you do not have access to");
     }
-
-    return status;
   }
 }
